Refactor apps code: extract shared utilities and eliminate duplication
This commit improves code organization and reduces duplication in the apps
commands by extracting shared YAML utility functions.

Changes:
- Create yaml_utils.go with shared utility functions:
  - camelToSnake: Convert camelCase strings to snake_case
  - yamlNodeToDynValue: Convert YAML nodes to dyn.Value
  - addBlankLinesBetweenTopLevelKeys: Format YAML files
- Move TestCamelToSnake to yaml_utils_test.go
- Remove duplicated functions from import.go
- Add .gitignore creation for AppKit templates in init.go
- Simplify .env template format (remove quotes around values)
- Remove unused bufio import from import.go
- Remove redundant comment about field ordering
