+++ cmd/apps/legacy-template/env.tmpl
@@ -1,7 +1,7 @@
-{{if .DatabricksHost}}DATABRICKS_HOST="{{.DatabricksHost}}"
-{{end}}MLFLOW_TRACKING_URI="{{.MlflowTrackingURI}}"
+{{if .DatabricksHost}}DATABRICKS_HOST={{.DatabricksHost}}
+{{end}}MLFLOW_TRACKING_URI={{.MlflowTrackingURI}}
 
-DATABRICKS_APP_NAME="{{.AppName}}"
+DATABRICKS_APP_NAME={{.AppName}}
 DATABRICKS_APP_PORT=8000
 FLASK_RUN_HOST=127.0.0.1
 FLASK_RUN_PORT=8000
@@ -15,5 +15,5 @@ STREAMLIT_SERVER_PORT=8000
 UVICORN_HOST=127.0.0.1
 UVICORN_PORT=8000
 
-{{range .AppYmlVars}}{{.Name}}="{{.Value}}"
+{{range .AppYmlVars}}{{.Name}}={{.Value}}
 {{end}}
